Update playground routers
